epub: Fix cover path for root=level OPF files
If the OPF file is in the root dir, g_path_get_dirname returns "."
instead of an empty string. This results in an invalid key for the cover
image, because it looks like "./OEBPS/cover_path.jpg" instead of
"OEBPS/cover_path.jpg" and then unzipping for that path fails.